Handwritten Letter Recognition
The study aims to develop a prediction model that recognizes handwritten English alphabet letters with the highest possible accuracy. In this context, a large and diverse dataset has been used. The dataset was obtained from the "A-Z Handwritten Alphabets in CSV Format" dataset available on Kaggle. This dataset provides a strong foundation in the field of handwritten letter recognition and is suitable for improving the model's accuracy.